add instruction all include header

+++ include/CPU.h
@@ -6,7 +6,7 @@
 #define AMC_CPU_H
 
 #include <cstdint>
-#include "instructions/MIPSInstruction.h"
+#include "instructions/all.h"
 #include "RAM.h"
 #include "InstructionFactory.h"
 #include "InstructionParser.h"
@@ -16,12 +16,12 @@ namespace AMC {
     public:
         static const uint32_t ALIGN_BYTES = 4U;
 
-        static const uint REGISTER_ZERO = 0;
-        static const uint REGISTER_V0 = 2;
-        static const uint REGISTER_A0 = 4;
-        static const uint REGISTER_T0 = 8;
-        static const uint REGISTER_GP = 28;
-        static const uint REGISTER_SP = 29;
+        static const uint8_t REGISTER_ZERO = 0U;
+        static const uint8_t REGISTER_V0 = 2U;
+        static const uint8_t REGISTER_A0 = 4U;
+        static const uint8_t REGISTER_T0 = 8U;
+        static const uint8_t REGISTER_GP = 28U;
+        static const uint8_t REGISTER_SP = 29U;
 
         CPU(uint32_t t_pc, uint64_t t_sp, uint64_t t_gp, RAM &t_ram);
         ~CPU();
